@@ -14,18 +14,18 @@ description: A new Flutter project.
 version: 1.0.0+1
 
 environment:
-  sdk: ">=2.1.0 <3.0.0"
+  sdk: ">=3.1.3 <4.0.0"
 
 dependencies:
-  sqflite: ^1.2.0
-  url_launcher: ^5.0.2
+  sqflite: ^2.3.0
+  url_launcher: 6.0.0
 
   flutter:
     sdk: flutter
 
   # The following adds the Cupertino Icons font to your application.
   # Use with the CupertinoIcons class for iOS style icons.
-  cupertino_icons: ^0.1.2
+  cupertino_icons: ^1.0.6
 
 dev_dependencies:
   flutter_test:
