@@ -1,12 +1,10 @@
 package com.example.agenda
 
 import androidx.annotation.NonNull;
-import io.flutter.embedding.android.FlutterActivity
 import io.flutter.embedding.engine.FlutterEngine
-import io.flutter.plugins.GeneratedPluginRegistrant
+
+import io.flutter.embedding.android.FlutterActivity;
 
 class MainActivity: FlutterActivity() {
-    override fun configureFlutterEngine(@NonNull flutterEngine: FlutterEngine) {
-        GeneratedPluginRegistrant.registerWith(flutterEngine);
-    }
+
 }
